Removed models directory
