+++ config.go
@@ -1,6 +1,12 @@
 package timeago
 
-import . "github.com/SerhiiCho/timeago/models"
+// Config is the main configurations for the package
+type Config struct {
+	// Default value is "ru"
+	Language string
+	// Default value is "Europe/Kiev"
+	Location string
+}
 
 var config = Config{
 	Language: "ru",
