--- models/config.go
@@ -1,9 +1,0 @@
-package models
-
-// Config is the main configurations for the package
-type Config struct {
-	// Default value is "ru"
-	Language string
-	// Default value is "Europe/Kiev"
-	Location string
-}
