--- models/lang.go
@@ -1,27 +1,0 @@
-package models
-
-type Lang struct {
-	Ago            string
-	Online         string
-	Second         string
-	Seconds        string
-	SecondsSpecial string
-	Minute         string
-	Minutes        string
-	MinutesSpecial string
-	Hour           string
-	Hours          string
-	HoursSpecial   string
-	Day            string
-	Days           string
-	DaysSpecial    string
-	Week           string
-	Weeks          string
-	WeeksSpecial   string
-	Month          string
-	Months         string
-	MonthsSpecial  string
-	Year           string
-	Years          string
-	YearsSpecial   string
-}
