--- models/rule.go
@@ -1,7 +1,0 @@
-package models
-
-type Rule struct {
-	Single  bool
-	Plural  bool
-	Special bool
-}
