+++ parser.go
@@ -4,15 +4,14 @@ import (
 	"encoding/json"
 	"log"
 
-	"github.com/SerhiiCho/timeago/models"
 	"github.com/SerhiiCho/timeago/utils"
 )
 
 // Parses json file and unmarshals result into a struct
-func parseNeededFile(fileName string) models.Lang {
+func parseNeededFile(fileName string) Lang {
 	content := utils.GetFileContent(fileName)
 
-	var result models.Lang
+	var result Lang
 
 	err := json.Unmarshal(content, &result)
