+++ tests/ru_test.go
@@ -4,8 +4,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/SerhiiCho/timeago"
-	"github.com/SerhiiCho/timeago/models"
+	. "github.com/SerhiiCho/timeago"
 	. "github.com/SerhiiCho/timeago/utils"
 )
 
@@ -60,12 +59,12 @@ func TestParseRu(t *testing.T) {
 
 	for _, tc := range cases {
 		t.Run("result for "+tc.date, func(test *testing.T) {
-			timeago.SetConfig(models.Config{
+			SetConfig(Config{
 				Language: "ru",
 				Location: "Europe/Kiev",
 			})
 
-			if res := timeago.Parse(tc.date); res != tc.result {
+			if res := Parse(tc.date); res != tc.result {
 				test.Errorf("Result must be %s, but got %s instead", tc.result, res)
 			}
 		})
@@ -134,11 +133,11 @@ func TestParseRuWithOnlineFlag(t *testing.T) {
 	for _, tc := range cases {
 		t.Run("result for "+tc.date, func(test *testing.T) {
 
-			timeago.SetConfig(models.Config{
+			SetConfig(Config{
 				Language: "ru",
 			})
 
-			if res := timeago.Parse(tc.date, "online"); res != tc.result {
+			if res := Parse(tc.date, "online"); res != tc.result {
 				test.Errorf("Result must be %s, but got %s instead", tc.result, res)
 			}
 		})
@@ -166,12 +165,12 @@ func TestParseRuWithSeconds(t *testing.T) {
 
 	for _, tc := range cases {
 		t.Run("result for "+tc.date, func(test *testing.T) {
-			timeago.SetConfig(models.Config{
+			SetConfig(Config{
 				Language: "ru",
 				Location: "Europe/Kiev",
 			})
 
-			if res := timeago.Parse(tc.date); res != tc.result[0] && res != tc.result[1] {
+			if res := Parse(tc.date); res != tc.result[0] && res != tc.result[1] {
 				test.Errorf("Result must be %s or %s, but got %s instead", tc.result[0], tc.result[1], res)
 			}
 		})
